web_app/requirements.txt
alembic==1.8.1
async-timeout==4.0.2
attrs==22.2.0
click==8.1.3
colorama==0.4.5
dominate==2.7.0
exceptiongroup==1.1.0
Faker==15.1.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Injector==0.14.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
iniconfig==1.1.1
injector==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
redis==4.5.1
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
